ck-builder: move TensorDescriptor and related types to tensor_descriptor.hpp
At this point the tensor descriptor stuff has become larger than the actual
tensor buffer stuff, so lets move it into a separate file.
